Simplify proxy updater again
